move rendering to its own file
